more options for dealing with proxies
Now a setting can be added to the global (pastebinc.conf) config file
or to an individual config file to bypass the proxy for all (or one)
provider(s).

Additionally, a flag has been added (-B) to avoid bypassing the proxy
manually if, for instance, you have a global setting to bypass it,
but do not want to bypass it for a single invocation.

Closes #1, Closes #2